+++ src/Concordium/Client/Commands.hs
@@ -64,7 +64,6 @@ data Backend =
   GRPC
     { grpcHost   :: !HostName
     , grpcPort   :: !PortNumber
-    , grpcAuthenticationToken :: !String
     , grpcTarget :: !(Maybe String)
     , grpcRetryNum :: !Int
     , grpcUseTls :: !Bool }
@@ -523,7 +522,7 @@ versionOption =
   infoOption (showVersion version) (hidden <> long "version" <> help "Show version.")
 
 backendParser :: Parser Backend
-backendParser = GRPC <$> hostParser <*> portParser <*> grpcAuthenticationTokenParser <*> targetParser <*> retryNumParser <*> tlsParser
+backendParser = GRPC <$> hostParser <*> portParser <*> targetParser <*> retryNumParser <*> tlsParser
 
 hostParser :: Parser HostName
 hostParser =
@@ -543,15 +542,6 @@ portParser =
      value 20000 <> -- default value to match the node default GRPC port
      showDefault <>
      help "Port where the gRPC server is listening.")
-
-grpcAuthenticationTokenParser :: Parser String
-grpcAuthenticationTokenParser =
-  strOption
-    (long "grpc-authentication-token" <>
-     metavar "GRPC-AUTHENTICATION-TOKEN" <>
-     value "rpcadmin" <> -- default value matches the default value for the node
-     showDefault <>
-     help "gRPC authentication token used to talk to the node")
 
 targetParser :: Parser (Maybe String)
 targetParser =