+++ src/Concordium/Client/GRPC2.hs
@@ -2686,7 +2686,6 @@ data GrpcConfig =
   GrpcConfig
     { host   :: !HostName
     , port   :: !PortNumber
-    , grpcAuthenticationToken :: !String
     -- Target node, i.e. "node-0" for use with grpc-proxy.eu.test.concordium.com against testnet
     , target :: !(Maybe String)
     -- |Number of times to __retry__ to establish a connection. Thus a value of
@@ -2757,11 +2756,10 @@ runClientWithCookies hds cfg comp = runStateT (runExceptT $ runReaderT (_runClie
 
 mkGrpcClient :: GrpcConfig -> Maybe LoggerMethod -> ClientIO EnvData
 mkGrpcClient config mLogger =
-  let auth = ("authentication", BS8.pack $ grpcAuthenticationToken config)
-      header =
+  let header =
         case target config of
-          Just t  -> [auth, ("target", BS8.pack t)]
-          Nothing -> [auth]
+          Just t  -> [("target", BS8.pack t)]
+          Nothing -> []
       cfg = (grpcClientConfigSimple (host config) (port config) (useTls config))
                  { _grpcClientConfigCompression = uncompressed
                  , _grpcClientConfigHeaders = header